conduit-axum: Move `BytesRequest` to main package
