+++ Cargo.toml
@@ -46,6 +46,7 @@ futures-channel = { version = "=0.3.25", default-features = false }
 futures-util = "=0.3.25"
 hex = "=0.4.3"
 http = "=0.2.8"
+http-body = "=0.4.5"
 hyper = { version = "=0.14.23", features = ["client", "http1"] }
 indexmap = { version = "=1.9.2", features = ["serde-1"] }
 indicatif = "=0.17.2"
