+++ conduit-axum/Cargo.toml
@@ -12,7 +12,6 @@ rust-version = "1.56.0"
 axum = "=0.6.2"
 hyper = { version = "=0.14.23", features = ["server", "stream"] }
 http = "=0.2.8"
-http-body = "=0.4.5"
 percent-encoding = "=2.2.0"
 sentry-core = { version = "=0.29.1", features = ["client"] }
 thiserror = "=1.0.38"
